@@ -11,40 +11,41 @@
     "format": "prettier --write ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^3.2.4",
-    "@sveltejs/adapter-cloudflare": "^4.7.2",
-    "@sveltejs/kit": "^2.5.25",
-    "@sveltejs/vite-plugin-svelte": "^3.1.2",
+    "@sveltejs/adapter-auto": "^3.3.1",
+    "@sveltejs/adapter-cloudflare": "^4.7.4",
+    "@sveltejs/kit": "^2.8.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0",
     "@types/three": "^0.168.0",
     "autoprefixer": "^10.4.20",
     "dotenv": "^16.4.5",
-    "postcss": "^8.4.43",
-    "svelte": "^4.2.19",
-    "svelte-check": "^3.8.6",
-    "tailwindcss": "^3.4.10",
-    "typescript": "^5.5.4",
-    "vite": "^5.4.2",
-    "wrangler": "^3.79.0"
+    "postcss": "^8.4.47",
+    "svelte": "^5.1.13",
+    "svelte-check": "^4.0.6",
+    "tailwindcss": "^3.4.14",
+    "typescript": "^5.6.3",
+    "vite": "^5.4.10",
+    "wrangler": "^3.86.0"
   },
   "type": "module",
   "dependencies": {
     "@cartridge/controller": "^0.3.46",
-    "@dojoengine/core": "^1.0.0-alpha.12",
-    "@dojoengine/create-burner": "^1.0.0-alpha.12",
+    "@dojoengine/core": "^1.0.0-alpha.30",
+    "@dojoengine/create-burner": "^1.0.0-alpha.30",
     "@dojoengine/recs": "^2.0.13",
-    "@dojoengine/state": "^1.0.0-alpha.15",
-    "@dojoengine/torii-client": "^1.0.0-alpha.15",
-    "@dojoengine/torii-wasm": "^1.0.0-alpha.15",
-    "@dojoengine/utils": "^1.0.0-alpha.15",
-    "@latticexyz/utils": "^2.2.0",
-    "@threlte/core": "^7.3.1",
-    "@threlte/extras": "^8.11.5",
+    "@dojoengine/state": "^1.0.0-alpha.30",
+    "@dojoengine/torii-client": "^1.0.0-alpha.30",
+    "@dojoengine/torii-wasm": "^1.0.0-alpha.30",
+    "@dojoengine/utils": "^1.0.0-alpha.30",
+
+    "@threlte/core": "^8.0.0-next.30",
+    "@threlte/extras": "^9.0.0-next.39",
+    
     "clsx": "^2.1.1",
     "lucide-svelte": "^0.445.0",
     "prettier": "^3.3.3",
-    "prettier-plugin-svelte": "^3.2.6",
+    "prettier-plugin-svelte": "^3.2.7",
     "starknet": "^6.11.0",
-    "tailwind-merge": "^2.5.2",
+    "tailwind-merge": "^2.5.4",
     "three": "^0.168.0",
     "three-inspect": "^0.7.2",
     "vite-plugin-top-level-await": "^1.4.4",
