@@ -1,29 +1,45 @@
 <script lang="ts">
   import { T } from '@threlte/core'
   import { onMount } from 'svelte'
-  import type { BulletCoords } from '$stores/coordsStores'
   import type { BufferGeometry, Points } from 'three'
   import { Color, Vector3 } from 'three'
+  import type { BulletWithPosition, Position } from '$lib/api/gameState'
 
-  export let bullet: BulletCoords
-  export let initialPosition: BulletCoords | undefined
-  export let trailSpacing: number = 0.01
-  export let compressionFactor: number = 1
+  interface Props {
+    bullet: BulletWithPosition
+    trailSpacing?: number
+    compressionFactor?: number
+  }
 
-  let geometry: BufferGeometry | undefined
-  let pointsRef: Points | undefined
+  let { bullet, trailSpacing = 0.01, compressionFactor = 1 }: Props = $props()
 
-  $: x = bullet.coords.x
-  $: y = bullet.coords.y
+  let geometryRef: any = $state()
+  let geometry: BufferGeometry | undefined = $derived(geometryRef?.geometry)
+  let pointsRef: Points | undefined = $state()
 
-  $: initialX = initialPosition?.coords.x ?? x
-  $: initialY = initialPosition?.coords.y ?? y
+  function normalizeCoords(coords: Position): Position {
+    return {
+      x: coords.x / 1000 - 50,
+      y: coords.y / 1000 - 50,
+    }
+  }
 
-  $: length = Math.sqrt((x - initialX) ** 2 + (y - initialY) ** 2)
-  $: count = Math.floor(length / trailSpacing)
+  let normalizedPos = $derived(normalizeCoords(bullet.position))
+
+  let x = $derived(normalizedPos.x)
+  let y = $derived(normalizedPos.y)
+
+  let normalizedInitial = $derived(normalizeCoords(bullet.shot_at))
+  let initialX = $derived(normalizedInitial.x)
+  let initialY = $derived(normalizedInitial.y)
+
+  let length = $derived(Math.sqrt((x - initialX) ** 2 + (y - initialY) ** 2))
+  let count = $derived(Math.floor(length / trailSpacing))
 
   // Calculate direction vector
-  $: direction = new Vector3(x - initialX, 0, y - initialY).normalize()
+  let direction = $derived(
+    new Vector3(x - initialX, 0, y - initialY).normalize()
+  )
 
   // Function to generate random red shades
   function getRandomRedShade() {
@@ -33,47 +49,51 @@
     return new Color(red / 255, green / 255, blue / 255)
   }
 
-  $: positions = (() => {
-    const posArray = new Float32Array(count * 3)
-    const colorArray = new Float32Array(count * 3)
-    for (let i = 0; i < count; i++) {
-      const t = i / (count - 1)
-      const compressedT = Math.pow(t, compressionFactor)
-      const px = initialX + (x - initialX) * compressedT
-      const py = initialY + (y - initialY) * compressedT
+  let positions = $derived(
+    (() => {
+      const posArray = new Float32Array(count * 3)
+      const colorArray = new Float32Array(count * 3)
+      for (let i = 0; i < count; i++) {
+        const t = i / (count - 1)
+        const compressedT = Math.pow(t, compressionFactor)
+        const px = initialX + (x - initialX) * compressedT
+        const py = initialY + (y - initialY) * compressedT
 
-      const randomFactor = 0.1
-      const vx = px + (Math.random() - 0.5) * randomFactor
-      const vy = 1 + (Math.random() - 0.5) * randomFactor
-      const vz = py + (Math.random() - 0.5) * randomFactor
+        const randomFactor = 0.1
+        const vx = px + (Math.random() - 0.5) * randomFactor
+        const vy = 1 + (Math.random() - 0.5) * randomFactor
+        const vz = py + (Math.random() - 0.5) * randomFactor
 
-      posArray[i * 3 + 0] = vx // x
-      posArray[i * 3 + 1] = vy // y
-      posArray[i * 3 + 2] = vz // z
+        posArray[i * 3 + 0] = vx // x
+        posArray[i * 3 + 1] = vy // y
+        posArray[i * 3 + 2] = vz // z
 
-      const color = getRandomRedShade()
-      colorArray[i * 3 + 0] = color.r
-      colorArray[i * 3 + 1] = color.g
-      colorArray[i * 3 + 2] = color.b
+        const color = getRandomRedShade()
+        colorArray[i * 3 + 0] = color.r
+        colorArray[i * 3 + 1] = color.g
+        colorArray[i * 3 + 2] = color.b
+      }
+      return { positions: posArray, colors: colorArray }
+    })()
+  )
+
+  $effect(() => {
+    if (geometry && positions) {
+      geometry.attributes.position.needsUpdate = true
+      geometry.computeBoundingSphere()
+
+      if (geometry.boundingSphere) {
+        const radius =
+          Math.max(Math.abs(x - initialX), Math.abs(y - initialY), 1) * 2
+        geometry.boundingSphere.radius = radius
+        geometry.boundingSphere.center.set(
+          (x + initialX) / 2,
+          1,
+          (y + initialY) / 2
+        )
+      }
     }
-    return { positions: posArray, colors: colorArray }
-  })()
-
-  $: if (geometry && positions) {
-    geometry.attributes.position.needsUpdate = true
-    geometry.computeBoundingSphere()
-
-    if (geometry.boundingSphere) {
-      const radius =
-        Math.max(Math.abs(x - initialX), Math.abs(y - initialY), 1) * 2
-      geometry.boundingSphere.radius = radius
-      geometry.boundingSphere.center.set(
-        (x + initialX) / 2,
-        1,
-        (y + initialY) / 2
-      )
-    }
-  }
+  })
 
   onMount(() => {
     if (pointsRef) {
@@ -85,17 +105,17 @@
 <T.Group>
   <!-- Bullet trail -->
   <T.Points bind:ref={pointsRef}>
-    <T.BufferGeometry bind:geometry>
+    <T.BufferGeometry bind:ref={geometryRef}>
       <T.BufferAttribute
         args={[positions.positions, 3]}
-        attach={(parent, self) => {
-          parent.setAttribute('position', self)
+        attach={({ parent, ref }) => {
+          ;(parent as any).setAttribute('position', ref)
         }}
       />
       <T.BufferAttribute
         args={[positions.colors, 3]}
-        attach={(parent, self) => {
-          parent.setAttribute('color', self)
+        attach={({ parent, ref }) => {
+          ;(parent as any).setAttribute('color', ref)
         }}
       />
     </T.BufferGeometry>
