@@ -2,18 +2,42 @@
   import { T } from '@threlte/core'
   import { PerspectiveCamera } from 'three'
   import PointerLock from './SplitScreen/PointerLock.svelte'
-  import { playerCharacterId } from '$stores/gameStores'
-  import { playerCharacterCoords } from '$stores/coordsStores'
+
   import { birdView } from '$stores/cameraStores'
   import { useThrelte } from '@threlte/core'
   import Hand from '../models/hand.svelte'
+  import getGame from '$lib/api/svelte/context'
+  import type { Position } from '$lib/api/gameState'
+  import { onDestroy } from 'svelte'
 
-  $: playerCoords = $playerCharacterCoords[$playerCharacterId]
+  let { currentCharacter } = getGame()
 
-  let {renderer} = useThrelte()
+  let playerCoords: Position | undefined = $state(undefined)
 
-  export let cameras: PerspectiveCamera[] = []
-  export let numCameras: number = 1
+  // Red: For some reason the object is not reactive if I do not subscribe manually
+  const unsubscribe = currentCharacter.subscribe(
+    (char) => (playerCoords = char?.coords)
+  )
+
+  onDestroy(unsubscribe)
+
+  function normalizeCoords(coords: Position): Position {
+    return {
+      x: coords.x / 1000 - 50,
+      y: coords.y / 1000 - 50,
+    }
+  }
+
+  let normalizedPlayerCoords = $derived(
+    normalizeCoords(playerCoords ?? { x: 0, y: 0 })
+  )
+
+  interface Props {
+    cameras?: PerspectiveCamera[]
+    numCameras?: number
+  }
+
+  let { cameras = $bindable([]), numCameras = 1 }: Props = $props()
 </script>
 
 {#if !$birdView}
@@ -22,10 +46,11 @@
 
 {#each Array(numCameras) as _, index}
   <T.PerspectiveCamera
-    position={[playerCoords.x, 1, playerCoords.y]}
-    on:create={({ ref }) => {
-      cameras[index] = ref
-      ref.lookAt(0, 1, 0)
+    position={[normalizedPlayerCoords.x, 1, normalizedPlayerCoords.y]}
+    oncreate={(obj) => {
+      console.log('plz')
+      cameras[index] = obj
+      obj.lookAt(0, 1, 0)
     }}
   >
     <Hand position={[0.1, -0.16, -0.6]} rotation={[0, Math.PI, 0]} />
