@@ -1,35 +1,79 @@
 <!--
 Auto-generated by: https://github.com/threlte/threlte/tree/main/packages/gltf
-Command: npx @threlte/gltf@2.0.3 container.glb --transform
+Command: npx @threlte/gltf@3.0.0-next.10 ./static/3d/container-transformed.glb -t --draco
 -->
 
 <script lang="ts">
-  import { Group, type Material } from 'three'
-  import { T, forwardEventHandlers } from '@threlte/core'
-  import { useGltf } from '@threlte/extras'
+  import type * as THREE from 'three'
 
-  export const ref = new Group()
+  import type { Snippet } from 'svelte'
+  import { T, type Props } from '@threlte/core'
+  import { useDraco, useGltf } from '@threlte/extras'
 
-  const gltf = useGltf('/3d/container-transformed.glb', { useDraco: true })
+  let {
+    fallback,
+    error,
+    children,
+    ref = $bindable(),
+    ...props
+  }: Props<THREE.Group> & {
+    ref?: THREE.Group
+    children?: Snippet<[{ ref: THREE.Group }]>
+    fallback?: Snippet
+    error?: Snippet<[{ error: Error }]>
+  } = $props()
 
-  const component = forwardEventHandlers()
-
-  gltf.then((gltf) => {
-    for (const [name, material] of Object.entries(gltf.materials) as [
-      string,
-      Material,
-    ][]) {
-      if (['CENTER GLASS ', 'GLASS BORDER', 'Material.008'].includes(name)) {
-        material.transparent = true
-        material.opacity = 0.8
-      }
+  type GLTFResult = {
+    nodes: {
+      mesh_0: THREE.Mesh
+      mesh_0_1: THREE.Mesh
+      mesh_0_2: THREE.Mesh
+      mesh_0_3: THREE.Mesh
+      mesh_0_4: THREE.Mesh
+      mesh_0_5: THREE.Mesh
+      mesh_0_6: THREE.Mesh
+      mesh_0_7: THREE.Mesh
+      mesh_0_8: THREE.Mesh
+      mesh_0_9: THREE.Mesh
+      mesh_0_10: THREE.Mesh
+      mesh_0_11: THREE.Mesh
+      mesh_0_12: THREE.Mesh
+      mesh_0_13: THREE.Mesh
+      mesh_0_14: THREE.Mesh
+      mesh_1: THREE.Mesh
+      mesh_1_1: THREE.Mesh
+      mesh_2: THREE.Mesh
+      mesh_2_1: THREE.Mesh
     }
+    materials: {
+      ['Material.007']: THREE.MeshPhysicalMaterial
+      ['Material.008']: THREE.MeshPhysicalMaterial
+      ['Material.011']: THREE.MeshStandardMaterial
+      WING: THREE.MeshStandardMaterial
+      ['ARC 2']: THREE.MeshStandardMaterial
+      ['ARC 1']: THREE.MeshStandardMaterial
+      ROMBO: THREE.MeshStandardMaterial
+      BEAMS: THREE.MeshStandardMaterial
+      ['CENTER TOP']: THREE.MeshStandardMaterial
+      ['GLASS BORDER']: THREE.MeshPhysicalMaterial
+      ['CENTER GLASS ']: THREE.MeshPhysicalMaterial
+      ['Material.010']: THREE.MeshStandardMaterial
+      ['Material.009']: THREE.MeshStandardMaterial
+      ['Material.006']: THREE.MeshStandardMaterial
+      ['Material.005']: THREE.MeshStandardMaterial
+      ['Material.001']: THREE.MeshStandardMaterial
+    }
+  }
+  const dracoLoader = useDraco()
+
+  const gltf = useGltf<GLTFResult>('/3d/container-transformed.glb', {
+    dracoLoader,
   })
 </script>
 
-<T is={ref} dispose={false} {...$$restProps} bind:this={$component}>
+<T.Group bind:ref dispose={false} {...props}>
   {#await gltf}
-    <slot name="fallback" />
+    {@render fallback?.()}
   {:then gltf}
     <T.Group rotation={[-Math.PI, 0, 0]}>
       <T.Mesh
@@ -113,9 +157,9 @@ Command: npx @threlte/gltf@2.0.3 container.glb --transform
         material={gltf.materials['Material.001']}
       />
     </T.Group>
-  {:catch error}
-    <slot name="error" {error} />
+  {:catch err}
+    {@render error?.({ error: err })}
   {/await}
 
-  <slot {ref} />
-</T>
+  {@render children?.({ ref })}
+</T.Group>
