@@ -1,23 +1,87 @@
 <!--
 Auto-generated by: https://github.com/threlte/threlte/tree/main/packages/gltf
-Command: npx @threlte/gltf@2.0.3 /3d"/hand.glb --transform
+Command: npx @threlte/gltf@3.0.0-next.10 ./static/3d/hand.glb --transform -t
 -->
 
-<script>
-  import { Group } from 'three'
-  import { T, forwardEventHandlers } from '@threlte/core'
-  import { useGltf } from '@threlte/extras'
+<script lang="ts">
+  import type * as THREE from 'three'
 
-  export const ref = new Group()
+  import type { Snippet } from 'svelte'
+  import { T, type Props } from '@threlte/core'
+  import { useDraco, useGltf } from '@threlte/extras'
 
-  const gltf = useGltf('/3d/hand-transformed.glb', { useDraco: true })
+  let {
+    fallback,
+    error,
+    children,
+    ref = $bindable(),
+    ...props
+  }: Props<THREE.Group> & {
+    ref?: THREE.Group
+    children?: Snippet<[{ ref: THREE.Group }]>
+    fallback?: Snippet
+    error?: Snippet<[{ error: Error }]>
+  } = $props()
 
-  const component = forwardEventHandlers()
+  type GLTFResult = {
+    nodes: {
+      pasted__pCylinder1: THREE.Mesh
+      pasted__pasted__pCylinder1: THREE.Mesh
+      pasted__pPlane5: THREE.Mesh
+      pasted__pasted__pPlane5: THREE.Mesh
+      pasted__pPlane5001: THREE.Mesh
+      pasted__pCylinder1001: THREE.Mesh
+      pasted__pCylinder2: THREE.Mesh
+      pasted__pPlane7: THREE.Mesh
+      pPlane1: THREE.Mesh
+      pPlane2: THREE.Mesh
+      pPlane3: THREE.Mesh
+      pPlane4: THREE.Mesh
+      pPipe1: THREE.Mesh
+      pPlane5: THREE.Mesh
+      pPlane6: THREE.Mesh
+      pCylinder1: THREE.Mesh
+      pCylinder2: THREE.Mesh
+      pPlane7: THREE.Mesh
+      pPlane8: THREE.Mesh
+      pPlane9: THREE.Mesh
+      pPipe2: THREE.Mesh
+      pCylinder3: THREE.Mesh
+      pPipe3: THREE.Mesh
+      pPlane10: THREE.Mesh
+      pSphere1: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F5c: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F5b: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F5a: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F4c: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F4b: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F4a: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F3c: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F3b: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F3a: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F2c: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F2b: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F2a: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F1b: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_F1a: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_UA: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_H2: THREE.Mesh
+      CSS_Arms_OBJ_Smoothed_H1: THREE.Mesh
+    }
+    materials: {
+      lambert1: THREE.MeshStandardMaterial
+      ['CSS_Arms_OBJ_Smoothed_:default2']: THREE.MeshStandardMaterial
+    }
+  }
+  const dracoLoader = useDraco()
+  const gltf = useGltf<GLTFResult>('/3d/hand-transformed.glb', {
+    dracoLoader,
+  })
 </script>
 
-<T is={ref} dispose={false} {...$$restProps} bind:this={$component}>
+<T.Group bind:ref dispose={false} {...props}>
   {#await gltf}
-    <slot name="fallback" />
+    {@render fallback?.()}
   {:then gltf}
     <T.Group rotation={[Math.PI / 2, 0, 0]} scale={0.01}>
       <T.Mesh
@@ -308,9 +372,9 @@ Command: npx @threlte/gltf@2.0.3 /3d"/hand.glb --transform
       rotation={[1.16, 0, 0]}
       scale={1.08}
     />
-  {:catch error}
-    <slot name="error" {error} />
+  {:catch err}
+    {@render error?.({ error: err })}
   {/await}
 
-  <slot {ref} />
-</T>
+  {@render children?.({ ref })}
+</T.Group>
