@@ -1,6 +1,8 @@
 <script lang="ts">
-  import { frameCounter } from '$src/stores/gameStores'
+  import getGame from '$lib/api/svelte/context'
   import { RECORDING_FRAME_LIMIT } from '$src/lib/consts'
+
+  const { frameCounter } = getGame()
 
   function getColor(frameCount: number) {
     let remaining = RECORDING_FRAME_LIMIT - frameCount
@@ -13,7 +15,9 @@
     return '#4caf50'
   }
 
-  $: barWidth = Math.max(0, 100 - ($frameCounter / RECORDING_FRAME_LIMIT) * 100)
+  let barWidth = $derived(
+    Math.max(0, 100 - ($frameCounter / RECORDING_FRAME_LIMIT) * 100)
+  )
 </script>
 
 <div class="step-bar-container">
