@@ -1,9 +1,16 @@
-<script>
+<script lang="ts">
   import { cn } from '$lib/css/cn'
   import Background from '$lib/ui/Background.svelte'
+  import type { Snippet } from 'svelte'
   import { slide } from 'svelte/transition'
 
-  export let color = ''
+  let {
+    color = '',
+    children,
+  }: {
+    color?: string
+    children?: Snippet
+  } = $props()
 </script>
 
 <div
@@ -14,7 +21,7 @@
     transition:slide
   >
     <div>
-      <slot></slot>
+      {@render children?.()}
     </div>
 
     <Background
