@@ -1,5 +1,5 @@
 
-// Generated by dojo-bindgen on Fri, 4 Oct 2024 22:01:20 +0000. Do not modify this file manually.
+// Generated by dojo-bindgen on Tue, 15 Oct 2024 11:08:47 +0000. Do not modify this file manually.
 // Import the necessary types from the recs SDK
 // generate again with `sozo build --typescript` 
 import { Account, byteArray } from "starknet";
@@ -14,22 +14,20 @@ export async function setupWorld(provider: DojoProvider) {
         const contract_name = "actions";
 
         
-        // Call the `move` system with the specified Account and calldata
-        const move = async (props: { account: Account, session_id: number, moves: models.TurnMove }) => {
+        // Call the `name` system with the specified Account and calldata
+        const name = async (props: { account: Account }) => {
             try {
                 return await provider.execute(
                     props.account,
                     {
                         contractName: contract_name,
-                        entrypoint: "move",
-                        calldata: [props.session_id,
-                props.moves.sub_moves,
-                    props.moves.shots],
+                        entrypoint: "name",
+                        calldata: [],
                     },
                     "octoguns"
                 );
             } catch (error) {
-                console.error("Error executing move:", error);
+                console.error("Error executing name:", error);
                 throw error;
             }
         };
@@ -55,8 +53,30 @@ export async function setupWorld(provider: DojoProvider) {
         };
             
 
+    
+        // Call the `move` system with the specified Account and calldata
+        const move = async (props: { account: Account, session_id: number, moves: models.TurnMove }) => {
+            try {
+                return await provider.execute(
+                    props.account,
+                    {
+                        contractName: contract_name,
+                        entrypoint: "move",
+                        calldata: [props.session_id,
+                props.moves.sub_moves,
+                    props.moves.shots],
+                    },
+                    "octoguns"
+                );
+            } catch (error) {
+                console.error("Error executing move:", error);
+                throw error;
+            }
+        };
+            
+
         return {
-            move, world
+            name, world, move
         };
     }
 
@@ -85,15 +105,37 @@ export async function setupWorld(provider: DojoProvider) {
             
 
     
+        // Call the `name` system with the specified Account and calldata
+        const name = async (props: { account: Account }) => {
+            try {
+                return await provider.execute(
+                    props.account,
+                    {
+                        contractName: contract_name,
+                        entrypoint: "name",
+                        calldata: [],
+                    },
+                    "octoguns"
+                );
+            } catch (error) {
+                console.error("Error executing name:", error);
+                throw error;
+            }
+        };
+            
+
+    
         // Call the `create` system with the specified Account and calldata
-        const create = async (props: { account: Account, objects: models.Map }) => {
+        const create = async (props: { account: Account, grid1: bigint, grid2: bigint, grid3: bigint }) => {
             try {
                 return await provider.execute(
                     props.account,
                     {
                         contractName: contract_name,
                         entrypoint: "create",
-                        calldata: [props.objects],
+                        calldata: [props.grid1,
+                props.grid2,
+                props.grid3],
                     },
                     "octoguns"
                 );
@@ -125,7 +167,7 @@ export async function setupWorld(provider: DojoProvider) {
             
 
         return {
-            world, create, default_map
+            world, name, create, default_map
         };
     }
 
@@ -173,8 +215,28 @@ export async function setupWorld(provider: DojoProvider) {
         };
             
 
+    
+        // Call the `name` system with the specified Account and calldata
+        const name = async (props: { account: Account }) => {
+            try {
+                return await provider.execute(
+                    props.account,
+                    {
+                        contractName: contract_name,
+                        entrypoint: "name",
+                        calldata: [],
+                    },
+                    "octoguns"
+                );
+            } catch (error) {
+                console.error("Error executing name:", error);
+                throw error;
+            }
+        };
+            
+
         return {
-            world, spawn
+            world, spawn, name
         };
     }
 
@@ -183,6 +245,26 @@ export async function setupWorld(provider: DojoProvider) {
         const contract_name = "start";
 
         
+        // Call the `name` system with the specified Account and calldata
+        const name = async (props: { account: Account }) => {
+            try {
+                return await provider.execute(
+                    props.account,
+                    {
+                        contractName: contract_name,
+                        entrypoint: "name",
+                        calldata: [],
+                    },
+                    "octoguns"
+                );
+            } catch (error) {
+                console.error("Error executing name:", error);
+                throw error;
+            }
+        };
+            
+
+    
         // Call the `world` system with the specified Account and calldata
         const world = async (props: { account: Account }) => {
             try {
@@ -277,7 +359,7 @@ export async function setupWorld(provider: DojoProvider) {
             
 
         return {
-            world, create, create_closed, join
+            name, world, create, create_closed, join
         };
     }
 
