@@ -1,5 +1,5 @@
 
-// Generated by dojo-bindgen on Fri, 4 Oct 2024 22:01:20 +0000. Do not modify this file manually.
+// Generated by dojo-bindgen on Tue, 15 Oct 2024 11:08:47 +0000. Do not modify this file manually.
 // Import the necessary types from the recs SDK
 // generate again with `sozo build --typescript` 
 import { defineComponent, Type as RecsType, type World } from "@dojoengine/recs";
@@ -16,16 +16,6 @@ export const LayoutDefinition = {
     value: RecsType.String
 };
         
-// Type definition for `octoguns::models::bullet::Bullet` struct
-export interface Bullet {
-    bullet_id: Number;
-    shot_step: Number;
-    shot_by: Number;
-    shot_at: Vec2;
-    velocity: IVec2;
-    
-}
-
 export const IVec2Definition = {
     x: RecsType.Number,
     y: RecsType.Number,
@@ -34,16 +24,26 @@ export const IVec2Definition = {
     
 };
 
+// Type definition for `octoguns::types::Vec2` struct
+export interface Vec2 {
+    x: Number;
+    y: Number;
+    
+}
 export const Vec2Definition = {
     x: RecsType.Number,
     y: RecsType.Number,
     
 };
 
-// Type definition for `octoguns::types::Vec2` struct
-export interface Vec2 {
-    x: Number;
-    y: Number;
+
+// Type definition for `octoguns::models::bullet::Bullet` struct
+export interface Bullet {
+    bullet_id: Number;
+    shot_step: Number;
+    shot_by: Number;
+    shot_at: Vec2;
+    velocity: IVec2;
     
 }
 export const BulletDefinition = {
@@ -89,8 +89,6 @@ export interface IVec2 {
     ydir: Boolean;
     
 }
-
-
 
 // Type definition for `octoguns::models::characters::CharacterModel` struct
 export interface CharacterModel {
@@ -144,14 +142,32 @@ export const GlobalDefinition = {
 // Type definition for `octoguns::models::map::Map` struct
 export interface Map {
     map_id: Number;
-    map_objects: Number[];
+    grid1: U256;
+    grid2: U256;
+    grid3: U256;
     
 }
-export const MapDefinition = {
-    map_id: RecsType.Number,
-    map_objects: RecsType.NumberArray,
+
+export const U256Definition = {
+    low: RecsType.BigInt,
+    high: RecsType.BigInt,
     
 };
+
+export const MapDefinition = {
+    map_id: RecsType.Number,
+    grid1: U256Definition,
+    grid2: U256Definition,
+    grid3: U256Definition,
+    
+};
+
+// Type definition for `core::integer::u256` struct
+export interface U256 {
+    low: BigInt;
+    high: BigInt;
+    
+}
 
 
 // Type definition for `octoguns::models::player::Player` struct
@@ -360,14 +376,16 @@ export function defineContractComponents(world: World) {
                 world,
                 {
                     map_id: RecsType.Number,
-                    map_objects: RecsType.NumberArray,
+                    grid1: U256Definition,
+                    grid2: U256Definition,
+                    grid3: U256Definition,
                 },
                 {
                     metadata: {
                         namespace: "octoguns",
                         name: "Map",
-                        types: ["u32", "array"],
-                        customTypes: [],
+                        types: ["u32"],
+                        customTypes: ["U256", "U256", "U256"],
                     },
                 }
             );
