@@ -9,15 +9,20 @@ import { Account } from 'starknet'
 import type { ArraySignatureType } from 'starknet'
 import { BurnerManager } from '@dojoengine/create-burner'
 import { getSyncEntities, getSyncEvents } from '@dojoengine/state'
+import { account } from '$src/stores/account'
 export type SetupResult = Awaited<ReturnType<typeof setup>>
 
-export async function setup(worldAddress: string | undefined, { ...config }: DojoConfig) {
+export async function setup(
+  worldAddress: string | undefined,
+  { ...config }: DojoConfig
+) {
   // torii client
   const toriiClient = await torii.createClient({
     rpcUrl: config.rpcUrl,
     toriiUrl: config.toriiUrl,
     relayUrl: '',
-    worldAddress: worldAddress ?? ''
+    worldAddress:
+      '0x190ca7e6b8c28576ae3616add6f7ff4cf454ee60460238226d8284ca77445e2',
   })
 
   // create contract components
@@ -29,33 +34,38 @@ export async function setup(worldAddress: string | undefined, { ...config }: Doj
   // create dojo provider
   const dojoProvider = new DojoProvider(config.manifest, config.rpcUrl)
 
-  const sync = await getSyncEntities(toriiClient, contractComponents as any, [])
+  const sync = await getSyncEntities(
+    toriiClient,
+    contractComponents as any,
+    undefined,
+    []
+  )
 
   // setup world
   const client = await setupWorld(dojoProvider)
-
   // create burner manager
-  // const burnerManager = new BurnerManager({
-  //   masterAccount: new Account(
-  //     {
-  //       nodeUrl: config.rpcUrl,
-  //     },
-  //     KATANA_PREFUNDED_ADDRESS,
-  //     KATANA_PREFUNDED_PRIVATE_KEY
-  //   ),
-  //   accountClassHash: config.accountClassHash,
-  //   rpcProvider: dojoProvider.provider,
-  //   feeTokenAddress: config.feeTokenAddress,
-  // })
+  const burnerManager = new BurnerManager({
+    masterAccount: new Account(
+      {
+        nodeUrl: config.rpcUrl,
+      },
+      config.masterAddress,
+      config.masterPrivateKey
+    ),
+    accountClassHash: config.accountClassHash,
+    rpcProvider: dojoProvider.provider,
+    feeTokenAddress: config.feeTokenAddress,
+  })
 
-  // try {
-  //   await burnerManager.init()
-  //   if (burnerManager.list().length === 0) {
-  //     await burnerManager.create()
-  //   }
-  // } catch (e) {
-  //   console.error(e)
-  // }
+  try {
+    console.log('Starting burner!')
+    await burnerManager.init()
+    if (burnerManager.list().length === 0) {
+      await burnerManager.create()
+    }
+  } catch (e) {
+    console.error('An error occurred while creating burner:', e)
+  }
 
   return {
     client,
@@ -66,7 +76,7 @@ export async function setup(worldAddress: string | undefined, { ...config }: Doj
     },
     config,
     dojoProvider,
-    // burnerManager,
+    burnerManager,
     toriiClient,
 
     torii,
