@@ -1,18 +1,12 @@
-import { get } from 'svelte/store';
-import manifest from '$manifests/deployment/manifest.json';
-import { createDojoConfig } from '@dojoengine/core';
-import { CONFIG } from '$stores/network'; 
+import { get } from 'svelte/store'
+import manifest from '$manifests/manifest_dev.json'
+import { createDojoConfig } from '@dojoengine/core'
+import { CONFIG } from '$stores/network'
 
-export function getDojoConfig() {
-  const configValue = get(CONFIG);
+export const dojoConfig = createDojoConfig({
+  toriiUrl: 'https://api.cartridge.gg/x/octoguns-alpha2/torii',
+  rpcUrl: 'https://api.cartridge.gg/x/octoguns-alpha2/katana',
+  manifest,
+})
 
-  const dojoConfig = createDojoConfig({
-    toriiUrl: configValue.PUBLIC_TORII_URL,
-    rpcUrl: configValue.PUBLIC_RPC_URL,
-    manifest,
-  });
-
-  const WORLD_ADDRESS = configValue.PUBLIC_WORLD_ADDRESS;
-
-  return { dojoConfig, WORLD_ADDRESS };
-}
+export type Config = typeof dojoConfig
