@@ -1,6 +1,6 @@
-export function load({ params }) {
-    return {
-      network: params.slug,
-    }
-  }
-  
+export function load({ url }) {
+  const firstSegment = url.pathname.split('/').filter(Boolean)[0]; // Extract the first path segment
+  return {
+      network: firstSegment,
+  };
+}
