@@ -1,86 +1,18 @@
 <script lang="ts">
-  import { dojoStore } from '$stores/dojoStore';
-  import { componentValueStore } from '$dojo/componentValueStore';
-  import GameList from '$lib/games/GameList.svelte';
-  import { type Entity } from '@dojoengine/recs';
-  import Button from '$lib/ui/Button.svelte';
-  import { cn } from '$lib/css/cn';
-  import { joinSession } from '$lib/game';
-  import { account } from '$stores/account';
-  import { env } from '$stores/network';
+  import GameList from '$lib/games/GameList.svelte'
+  import { type Entity } from '@dojoengine/recs'
+  import Button from '$lib/ui/Button.svelte'
+  import { cn } from '$lib/css/cn'
+  import { env } from '$stores/network'
+  import { openSessions } from '$lib/api/sessions'
+  import { openSessionMetas } from '$lib/api/sessionMeta'
+  import type { Session, SessionMeta } from '$src/dojo/models.gen'
 
-  type Session = {
-    value: any;
-    isYourTurn: boolean;
-    isStarted: boolean;
-    isFinished: boolean;
-    enemy: string;
-    username?: string;
-  };
+  // TODO: Migrate this to directly using stores
 
+  let availableSessions: Session[] | null = $derived($openSessions)
+  let availableSessionMetas: SessionMeta[] | null = $derived($openSessionMetas)
 
-  let sessions: Session[] = [];
-  let playerEntity: Entity;
-
-  $: ({ clientComponents, torii } = $dojoStore as any);
-
-  $: globalEntity = torii.poseidonHash([BigInt(0).toString()]);
-
-  $: if ($account) playerEntity = torii.poseidonHash([$account.address]);
-
-  $: global = componentValueStore(clientComponents.Global, globalEntity);
-  $: player = componentValueStore(clientComponents.Player, playerEntity);
-
-
-  $: if ($global) {
-    let currentSessions = [];
-    let availableSessions = [];
-
-    if ($player) {
-      currentSessions = $player.games.map((game: { value: any }) => game.value);
-
-      const playerGamesSet = new Set(currentSessions);
-
-      availableSessions = $global.pending_sessions.filter(
-        (session: { value: any }) => !playerGamesSet.has(session.value)
-      );
-    } else {
-      availableSessions = $global.pending_sessions;
-    }
-
-    sessions = [];
-
-    for (const session of availableSessions) {
-      const sessionEntity = torii.poseidonHash([BigInt(session.value).toString()]);
-
-      if (sessionEntity) {
-        const sessionDataStore = componentValueStore(clientComponents.Session, sessionEntity);
-        const sessionMetaDataStore = componentValueStore(clientComponents.SessionMeta, sessionEntity);
-
-        sessionDataStore.subscribe(async (data) => {
-          if (data) {
-            const enemyAddress = `0x${BigInt(data.player1).toString(16)}`;
-            const newSession: Session = {
-              value: session.value,
-              isYourTurn: false,
-              isStarted: false,
-              isFinished: data.state === 3,
-              enemy: enemyAddress,
-              username: undefined,
-            };
-
-            sessionMetaDataStore.subscribe((metaData) => {
-              if (metaData) {
-                newSession.isStarted = metaData.p1_character !== 0;
-              }
-            });
-
-            sessions = [...sessions, newSession];
-          }
-        });
-      }
-    }
-  }
 </script>
 
 <div class={cn('flex flex-col h-full')}>
@@ -89,12 +21,17 @@
     <span class="flex-grow"></span>
     <Button href={`/${$env}/client/games/create`}>+ New Game</Button>
   </div>
-  <div class={cn('flex flex-col', { 'justify-center': sessions.length === 0 })}>
-    {#if sessions && sessions.length > 0}
+  <div
+    class={cn('flex flex-col', {
+      'justify-center': !availableSessions,
+    })}
+  >
+    {#if availableSessions && availableSessions.length > 0}
       <h1 class="text-xl ml-5 mb-3 font-bold">Games available</h1>
       <GameList
-        availableSessions={sessions}
-        on:select={(session) => joinSession(session.detail)}
+        {availableSessions}
+        {availableSessionMetas}
+        contractCall={true}
       />
     {:else}
       <div class="self-center align-middle flex flex-col gap-2">
